refactor: consolidate getDomain and isLocalhostOrIPAddress into url-utils
