+++ app/util/url-utils.ts
@@ -1,5 +1,26 @@
+import ipRegex from 'ip-regex';
 import psl from 'psl';
-import isLocalhostOrIPAddress from './isLocalhostOrIPAddress';
+
+/**
+ * Check if a hostname is localhost or an IP address (v4 or v6). Public RPC
+ * providers use domain names, not raw IP addresses, so a `true` result means
+ * the host should not be reduced to an eTLD+1.
+ *
+ * @param hostname - The hostname to check.
+ * @returns True if the hostname is localhost or an IP address.
+ */
+export function isLocalhostOrIPAddress(hostname: string): boolean {
+  const lowerHostname = hostname.toLowerCase();
+
+  if (lowerHostname === 'localhost') {
+    return true;
+  }
+
+  // Remove brackets from IPv6 addresses for testing (e.g., [::1] -> ::1)
+  const hostnameWithoutBrackets = lowerHostname.replace(/^\[|\]$/gu, '');
+
+  return ipRegex({ exact: true }).test(hostnameWithoutBrackets);
+}
 
 /**
  * Registrable domain (eTLD+1) for a URL, computed via the Public Suffix List
@@ -14,7 +35,7 @@ import isLocalhostOrIPAddress from './isLocalhostOrIPAddress';
  * @param urlString - The URL to extract a domain from.
  * @returns The domain, or null if the URL is invalid.
  */
-function getDomain(urlString: string): string | null {
+export function getDomain(urlString: string): string | null {
   let url: URL;
   try {
     url = new URL(urlString);
@@ -30,5 +51,3 @@ function getDomain(urlString: string): string | null {
 
   return psl.get(hostname) ?? hostname;
 }
-
-export default getDomain;
